refactor: PR review comments
