+++ posthog/models/web_preaggregated/test_web_pre_aggregated_timezones.py
@@ -493,47 +493,6 @@ class TestTimezonePreAggregatedIntegration(WebAnalyticsPreAggregatedTestBase, Fl
         finally:
             india_team.delete()
 
-    def _add_extra_timezone_boundary_events(self, team):
-        """Add events that extend well into Jan 16 to cover timezone boundaries"""
-        # Add events throughout Jan 16 to ensure PST timezone coverage
-        additional_events = [
-            # Jan 16 early morning (covers PST evening of Jan 15)
-            ("2024-01-16T02:00:00Z", "Chrome", "boundary_user_1", "jan16_02h"),
-            ("2024-01-16T04:00:00Z", "Safari", "boundary_user_2", "jan16_04h"),
-            ("2024-01-16T06:00:00Z", "Firefox", "boundary_user_3", "jan16_06h"),
-            ("2024-01-16T08:00:00Z", "Edge", "boundary_user_4", "jan16_08h"),
-            ("2024-01-16T10:00:00Z", "Chrome", "boundary_user_5", "jan16_10h"),
-            ("2024-01-16T12:00:00Z", "Safari", "boundary_user_6", "jan16_12h"),
-            ("2024-01-16T14:00:00Z", "Firefox", "boundary_user_7", "jan16_14h"),
-            ("2024-01-16T16:00:00Z", "Edge", "boundary_user_8", "jan16_16h"),
-        ]
-
-        sessions = [str(uuid7("2024-01-16")) for _ in range(len(additional_events))]
-
-        # Create additional users
-        for i in range(len(additional_events)):
-            _create_person(team_id=team.pk, distinct_ids=[f"boundary_user_{i+1}"])
-
-        # Create additional events
-        for i, (timestamp, browser, user_id, label) in enumerate(additional_events):
-            _create_event(
-                team=team,
-                event="$pageview",
-                distinct_id=user_id,
-                timestamp=timestamp,
-                properties={
-                    "$session_id": sessions[i],
-                    "$current_url": f"https://example.com/{label}",
-                    "$pathname": f"/{label}",
-                    "$device_type": "Desktop",
-                    "$browser": browser,
-                    "$host": "example.com",
-                    "test_label": label,
-                },
-            )
-
-        flush_persons_and_events()
-
     def test_can_use_date_range_timezone_integration(self):
         """Test can_use_date_range function works correctly across different timezones with real data"""
         utc_team = self._create_timezone_team("UTC")
@@ -715,91 +674,6 @@ class TestCanUseDateRangeTimezones(WebAnalyticsPreAggregatedTestBase):
         return Team.objects.create(
             organization=self.organization, name=f"Team {timezone_name.replace('/', '_')}", timezone=timezone_name
         )
-
-    def _add_extra_timezone_boundary_events(self, team):
-        """Add events that extend well into Jan 16 to cover timezone boundaries"""
-        # Add events throughout Jan 16 to ensure PST timezone coverage
-        additional_events = [
-            # Jan 16 early morning (covers PST evening of Jan 15)
-            ("2024-01-16T02:00:00Z", "Chrome", "boundary_user_1", "jan16_02h"),
-            ("2024-01-16T04:00:00Z", "Safari", "boundary_user_2", "jan16_04h"),
-            ("2024-01-16T06:00:00Z", "Firefox", "boundary_user_3", "jan16_06h"),
-            ("2024-01-16T08:00:00Z", "Edge", "boundary_user_4", "jan16_08h"),
-            ("2024-01-16T10:00:00Z", "Chrome", "boundary_user_5", "jan16_10h"),
-            ("2024-01-16T12:00:00Z", "Safari", "boundary_user_6", "jan16_12h"),
-            ("2024-01-16T14:00:00Z", "Firefox", "boundary_user_7", "jan16_14h"),
-            ("2024-01-16T16:00:00Z", "Edge", "boundary_user_8", "jan16_16h"),
-        ]
-
-        sessions = [str(uuid7("2024-01-16")) for _ in range(len(additional_events))]
-
-        # Create additional users
-        for i in range(len(additional_events)):
-            _create_person(team_id=team.pk, distinct_ids=[f"boundary_user_{i+1}"])
-
-        # Create additional events
-        for i, (timestamp, browser, user_id, label) in enumerate(additional_events):
-            _create_event(
-                team=team,
-                event="$pageview",
-                distinct_id=user_id,
-                timestamp=timestamp,
-                properties={
-                    "$session_id": sessions[i],
-                    "$current_url": f"https://example.com/{label}",
-                    "$pathname": f"/{label}",
-                    "$device_type": "Desktop",
-                    "$browser": browser,
-                    "$host": "example.com",
-                    "test_label": label,
-                },
-            )
-
-        flush_persons_and_events()
-
-    def _setup_cross_timezone_test_data(self, team):
-        # Create events across multiple days with timezone-aware timestamps
-        events = [
-            # Day 1: Jan 14 - Events that will cross timezone boundaries
-            ("2024-01-14T06:00:00Z", "Chrome", "user_1", "early_utc"),
-            ("2024-01-14T12:00:00Z", "Safari", "user_2", "midday_utc"),
-            ("2024-01-14T18:00:00Z", "Firefox", "user_3", "evening_utc"),
-            ("2024-01-14T23:30:00Z", "Edge", "user_4", "late_utc"),
-            # Day 2: Jan 15 - Main test day
-            ("2024-01-15T03:00:00Z", "Chrome", "user_5", "early_jan15"),
-            ("2024-01-15T09:00:00Z", "Safari", "user_6", "morning_jan15"),
-            ("2024-01-15T15:00:00Z", "Firefox", "user_7", "afternoon_jan15"),
-            ("2024-01-15T21:00:00Z", "Edge", "user_8", "night_jan15"),
-            # Day 3: Jan 16 - Events for boundary testing
-            ("2024-01-16T02:00:00Z", "Chrome", "user_9", "jan16_early"),
-            ("2024-01-16T14:00:00Z", "Safari", "user_10", "jan16_afternoon"),
-        ]
-
-        sessions = [str(uuid7("2024-01-15")) for _ in range(len(events))]
-
-        # Create users
-        for i in range(len(events)):
-            _create_person(team_id=team.pk, distinct_ids=[f"user_{i+1}"])
-
-        # Create events with properties
-        for i, (timestamp, browser, user_id, label) in enumerate(events):
-            _create_event(
-                team=team,
-                event="$pageview",
-                distinct_id=user_id,
-                timestamp=timestamp,
-                properties={
-                    "$session_id": sessions[i],
-                    "$current_url": f"https://example.com/{label}",
-                    "$pathname": f"/{label}",
-                    "$device_type": "Desktop",
-                    "$browser": browser,
-                    "$host": "example.com",
-                    "test_label": label,
-                },
-            )
-
-        flush_persons_and_events()
 
     def _populate_preaggregated_tables(self, team):
         stats_insert = WEB_STATS_INSERT_SQL(
